refactor: HistoryView
