+++ Moa/Moa/Domain/Entity/History/Calendar/CalendarLabelStyle.swift
@@ -1,0 +1,10 @@
+//
+//  CalendarLabelStyle.swift
+//  Moa
+//
+//  Created by 정도현 on 3/17/26.
+//
+
+import Foundation
+
+enum CalendarLabelStyle: Equatable { case payday, vacation }
