+++ Moa/Moa/Domain/Entity/History/Calendar/HistoryError.swift
@@ -1,0 +1,13 @@
+//
+//  HistoryError.swift
+//  Moa
+//
+//  Created by 정도현 on 3/17/26.
+//
+
+import Foundation
+
+enum HistoryError: Equatable {
+    case network
+    case dataCorrupted
+}
