+++ Moa/Moa/Domain/Entity/History/Weekday.swift
